requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.26.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
pytest==8.0.0
